requirements.txt
flask==2.3.3
flask-cors==4.0.0
openai==1.3.0
requests==2.31.0
python-dotenv==1.0.0
rapidfuzz==3.5.2
tiktoken==0.5.2
pathlib